+++ projects/core/src/attach.ts
@@ -1,4 +1,5 @@
 import {ChangeDetectorRef, ElementRef, inject, ProviderToken, ViewRef} from "@angular/core";
+import {decorateFactory} from "./core";
 import {getMeta, selector, setMeta} from "./metadata";
 import {EVENTS} from "./providers";
 import {addDep, track} from "./proxy";
@@ -33,3 +34,14 @@ export function attach<T extends {}>(token: ProviderToken<T>, directive?: any, k
    addDep(obj, "check", value)
    return track(value)
 }
+
+export function Attach(token: ProviderToken<any>) {
+   return function (target: any, key: string) {
+      decorateFactory(target.constructor, function (target: any, factory: Function, ...args: any[]) {
+         const instance = factory(...args)
+         attach(token, instance, key)
+         return instance
+      })
+   }
+}
+
