+++ projects/core/src/decorators.ts
@@ -1,15 +1,13 @@
-import {ProviderToken} from "@angular/core";
-import {attach} from "./attach";
-import {ActionMetadata, Metadata, Phase, SelectMetadata, StatusMetadata} from "./interfaces";
 import {
-   setup,
    decorateActions,
    decorateChanges,
    decorateCheck,
    decorateDestroy,
    decorateFactory,
-   decorateSelectors
+   decorateSelectors,
+   setup
 } from "./core";
+import {ActionMetadata, Metadata, Phase, SelectMetadata, StatusMetadata} from "./interfaces";
 import {action, caught, getStatuses, selector, setMeta, status} from "./metadata";
 
 const defaults = {track: true, immediate: true}
@@ -47,14 +45,3 @@ export const Layout = createDecorator<ActionMetadata>(action, {...defaults, phas
 export const Select = createDecorator<SelectMetadata>(selector)
 export const Caught = createDecorator(caught)
 export const Status = createDecorator<{ action?: string }>(status)
-
-export function Attach(token: ProviderToken<any>) {
-   return function (target: any, key: string) {
-      decorateFactory(target.constructor, function (target: any, factory: Function, ...args: any[]) {
-         const instance = factory(...args)
-         attach(token, instance, key)
-         return instance
-      })
-   }
-}
-
