+++ projects/core/src/public-api.ts
@@ -13,11 +13,11 @@ export {
 } from "./proxy"
 export {dispatch} from "./dispatch";
 export {useOperator, useConcat, useExhaust, useMerge, useSwitch, addTeardown, useChanges} from "./hooks"
-export {attach} from "./attach"
+export {attach, Attach} from "./attach"
 export {TemplateProvider} from "./template-provider"
 export {loadEffect} from "./load-effect"
 export {Transition, noopTransition} from "./transition"
 export {EVENTS, ACTION} from "./providers";
 export {configureStore, events} from "./utils";
-export {Store, Status, Caught, Select, Layout, Before, Invoke, Action, Attach} from "./decorators";
+export {Store, Status, Caught, Select, Layout, Before, Invoke, Action} from "./decorators";
 export { select, selectStore, Selector } from "./select"
